+++ example/hello/src/hello.dart
@@ -18,13 +18,12 @@ import 'package:dart_web_toolkit/validation.dart' as validation;
 import 'package:dart_web_toolkit/uibinder.dart';
 
 void main() {
-  
+  dart_html.query("#loading").remove();
+  //
   ui.RootPanel.get().add(new TestPanel());
 }
 
 class TestPanel extends ui.Composite {
-  
-  UiBinder<ui.Widget, TestPanel> binder = new UiBinder<ui.Widget, TestPanel>();
   
   @UiField
   ui.Label message;
@@ -33,21 +32,24 @@ class TestPanel extends ui.Composite {
   ui.Button recalculateFleet;
   
   TestPanel() {
+
+    UiBinder<ui.Widget, TestPanel> binder = new UiBinder<ui.Widget, TestPanel>();
     
     // load the template e.g. from service ...
     binder.viewTemplate = "<div class='dyn-panel'><div ui:field='message' class='dyn-message'>Status messages are shown here</div><button ui:field='recalculateFleet'>Recalculate fleet</button></div>";
 
     // known from GWT UiBinder - initialize
-    binder.createAndBindUi(this);
-//    initWidget(binder.createAndBindUi(this));
+    initWidget(binder.createAndBindUi(this));
     
-//    recalculateFleet.addClickHandler(new event.ClickHandlerAdapter((event.ClickEvent evt){
-//      message.text = "Clicked";
-//    }));
+    recalculateFleet.addClickHandler(new event.ClickHandlerAdapter((event.ClickEvent evt){
+      message.text = "Clicked";
+    }));
   }
 }
 
-
+// 
+// <input id="new-todo" placeholder="What needs to be done?" autofocus>
+//
 void main_97() {
   dart_html.query("#loading").remove();
   // Create a panel to layout the widgets